Created authentication, authorization, and middleware.

+++ backend/src/controllers/post.js
@@ -1,4 +1,5 @@
 const Post = require("../models/post");
+const User = require("../models/user");
 
 const createPost = async (req, res) => {
   const { title, description } = req.body;
@@ -7,7 +8,15 @@ const createPost = async (req, res) => {
     const post = await Post.create({
       title,
       description,
+      author: req.user.userId,
     });
+
+    const user = await User.findById(req.user.userId);
+
+    if (user) {
+      user.posts.push(post._id);
+      await user.save();
+    }
 
     res.status(201).json(post);
   } catch (error) {
@@ -17,7 +26,7 @@ const createPost = async (req, res) => {
 
 const getAllPosts = async (req, res) => {
   try {
-    const posts = await Post.find();
+    const posts = await Post.find().populate("author", "username");
     res.status(200).json(posts);
   } catch (error) {
     res.status(400).json({ error: error.message });
@@ -28,7 +37,10 @@ const getPost = async (req, res) => {
   const { id } = req.params;
 
   try {
-    const post = await Post.findById({ _id: id });
+    const post = await Post.findById({ _id: id }).populate(
+      "author",
+      "username"
+    );
 
     if (!post) {
       return res.status(404).json({ error: "No post found." });
@@ -43,11 +55,20 @@ const deletePost = async (req, res) => {
   const { id } = req.params;
 
   try {
-    const post = await Post.findByIdAndDelete({ _id: id });
+    const post = await Post.findById({ _id: id });
 
     if (!post) {
       return res.status(404).json({ error: "No post found." });
     }
+
+    if (post.author.toString() !== req.user.userId) {
+      return res
+        .status(403)
+        .json({ error: "Not authorized to delete this post." });
+    }
+
+    await Post.findByIdAndDelete({ _id: id });
+
     res.status(200).json({ message: "Post successfully has been removed." });
   } catch (error) {
     res.status(400).json({ error: error.message });
@@ -58,16 +79,40 @@ const updatePost = async (req, res) => {
   const { id } = req.params;
 
   try {
-    const post = await Post.findByIdAndUpdate(
-      { _id: id },
-      { ...req.body },
-      { new: true, runValidators: true }
-    );
+    const post = await Post.findById({ _id: id });
 
     if (!post) {
       return res.status(404).json({ error: "No post found." });
     }
-    res.status(200).json({ message: "Post has been successfully updated." });
+
+    if (post.author.toString() !== req.user.userId) {
+      return res
+        .status(403)
+        .json({ error: "Not authorized to delete this post." });
+    }
+
+    const updatedPost = await Post.findByIdAndUpdate(
+      id,
+      { ...req.body },
+      { new: true, runValidators: true }
+    ).populate("author", "username");
+
+    res.status(200).json({
+      message: "Post has been successfully updated.",
+      post: updatedPost,
+    });
+  } catch (error) {
+    res.status(400).json({ error: error.message });
+  }
+};
+
+const getUserPosts = async (req, res) => {
+  try {
+    const posts = await Post.find({ author: req.user.userId }).populate(
+      "author",
+      "username"
+    );
+    res.status(200).json(posts);
   } catch (error) {
     res.status(400).json({ error: error.message });
   }
@@ -79,4 +124,5 @@ module.exports = {
   getPost,
   deletePost,
   updatePost,
+  getUserPosts,
 };
